+++ build_daemon/lib/daemon.dart
@@ -9,29 +9,12 @@ import 'package:built_value/serializer.dart';
 import 'package:pedantic/pedantic.dart';
 import 'package:watcher/watcher.dart';
 
-import '../constants.dart';
-import '../daemon_builder.dart';
-import '../data/build_target.dart';
-import 'file_wait.dart';
-import 'server.dart';
-
-/// Returns the current version of the running build daemon.
-///
-/// Null if one isn't running.
-Future<String> runningVersion(String workingDirectory) async {
-  var versionFile = File(versionFilePath(workingDirectory));
-  if (!await waitForFile(versionFile)) return null;
-  return versionFile.readAsStringSync();
-}
-
-/// Returns the current options of the running build daemon.
-///
-/// Null if one isn't running.
-Future<Set<String>> currentOptions(String workingDirectory) async {
-  var optionsFile = File(optionsFilePath(workingDirectory));
-  if (!await waitForFile(optionsFile)) return Set();
-  return optionsFile.readAsLinesSync().toSet();
-}
+import 'change_provider.dart';
+import 'constants.dart';
+import 'daemon_builder.dart';
+import 'data/build_target.dart';
+import 'src/file_wait.dart';
+import 'src/server.dart';
 
 /// The long running daemon process.
 ///
@@ -42,56 +25,68 @@ Future<Set<String>> currentOptions(String workingDirectory) async {
 /// notification.
 class Daemon {
   final String _workingDirectory;
-
+  final RandomAccessFile _lock;
   final _doneCompleter = Completer();
 
   Server _server;
   StreamSubscription _sub;
-  RandomAccessFile _lock;
 
-  Daemon(this._workingDirectory);
+  Daemon(String workingDirectory)
+      : _workingDirectory = workingDirectory,
+        _lock = _tryGetLock(workingDirectory);
 
   Future<void> get onDone => _doneCompleter.future;
 
   Future<void> stop({String message, int failureType}) =>
       _server.stop(message: message, failureType: failureType);
 
-  /// Starts the daemon.
+  bool get hasLock => _lock != null;
+
+  /// Returns the current version of the running build daemon.
   ///
-  /// [changes] is a stream of lists of file changes. If multiple files change
-  /// together then they should be sent on this stream in the same list.
-  /// Otherwise, at least two builds will be triggered.
-  Future<void> start(Set<String> options, DaemonBuilder builder,
-      Stream<List<WatchEvent>> changes,
-      {Serializers serializersOverride,
-      bool Function(BuildTarget, Iterable<WatchEvent>) shouldBuild,
-      Duration timeout = const Duration(seconds: 30)}) async {
-    if (_server != null) return;
+  /// Null if one isn't running.
+  Future<String> runningVersion() async {
+    var versionFile = File(versionFilePath(_workingDirectory));
+    if (!await waitForFile(versionFile)) return null;
+    return versionFile.readAsStringSync();
+  }
+
+  /// Returns the current options of the running build daemon.
+  ///
+  /// Null if one isn't running.
+  Future<Set<String>> currentOptions() async {
+    var optionsFile = File(optionsFilePath(_workingDirectory));
+    if (!await waitForFile(optionsFile)) return Set();
+    return optionsFile.readAsLinesSync().toSet();
+  }
+
+  Future<void> start(
+    Set<String> options,
+    DaemonBuilder builder,
+    ChangeProvider changeProvider, {
+    Serializers serializersOverride,
+    bool Function(BuildTarget, Iterable<WatchEvent>) shouldBuild,
+    Duration timeout = const Duration(seconds: 30),
+  }) async {
+    if (_server != null || _lock == null) return;
     _handleGracefulExit();
 
     _createVersionFile();
     _createOptionsFile(options);
 
-    _server = Server(builder, changes, timeout,
-        serializersOverride: serializersOverride, shouldBuild: shouldBuild);
+    _server = Server(
+      builder,
+      timeout,
+      changeProvider,
+      serializersOverride: serializersOverride,
+      shouldBuild: shouldBuild,
+    );
     var port = await _server.listen();
     _createPortFile(port);
 
     unawaited(_server.onDone.then((_) async {
       await _cleanUp();
     }));
-  }
-
-  bool tryGetLock() {
-    try {
-      _createDaemonWorkspace();
-      _lock = File(lockFilePath(_workingDirectory))
-          .openSync(mode: FileMode.write)
-            ..lockSync();
-      return true;
-    } on FileSystemException {
-      return false;
-    }
   }
 
   Future<void> _cleanUp() async {
@@ -104,14 +99,6 @@ class Daemon {
       workspace.deleteSync(recursive: true);
     }
     if (!_doneCompleter.isCompleted) _doneCompleter.complete();
-  }
-
-  void _createDaemonWorkspace() {
-    try {
-      Directory(daemonWorkspace(_workingDirectory)).createSync(recursive: true);
-    } catch (e) {
-      throw Exception('Unable to create daemon workspace: $e');
-    }
   }
 
   void _createPortFile(int port) =>
@@ -133,5 +120,25 @@ class Daemon {
         if (cancelCount > 1) exit(1);
       }
     });
+  }
+}
+
+RandomAccessFile _tryGetLock(String workingDirectory) {
+  try {
+    _createDaemonWorkspace(workingDirectory);
+    var lock = File(lockFilePath(workingDirectory))
+        .openSync(mode: FileMode.write)
+          ..lockSync();
+    return lock;
+  } on FileSystemException {
+    return null;
+  }
+}
+
+void _createDaemonWorkspace(String workingDirectory) {
+  try {
+    Directory(daemonWorkspace(workingDirectory)).createSync(recursive: true);
+  } catch (e) {
+    throw Exception('Unable to create daemon workspace: $e');
   }
 }